host-gmp: Update to 5.0.4
The reason for the upgrade is that the 5.0.1 version fails to build on a
Intel i5 with Mac OS.

gmp includes config.guess which compiles some assembly to detect the
host CPU. Since the config.guess of 5.0.1 is too old, it uses "uname -p"
which unfortunately returns "i386" on Mac OS even if it's 64 bit. That
causes gmp to be compiled with -m32 which works, but then the 64 bit gcc
cannot link against gmp. (I had a similar problem in a mis-configured
Linux VM [KVM and libvirt setup] recently, so it's not really
Mac-specific.) The new gmp prints coreisbr-apple-darwin11.3.0 which is
detected as 64 bit and everything works.

Apart from that, newer is better. :)

Signed-off-by: Bernhard Walle <bernhard@bwalle.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>
(cherry picked from commit d2f28a85553a2bf81f52ce83668304754f6b26e8)

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>