gcc: aarch64: fix multi-os-directory
The multi-os-directory, which can be printed with

	aarch64-v8a-linux-gnu-gcc -print-multi-os-directory

is ../lib for aarch64, but must be . to ensure a correct search path for
libraries that use multi-os-directory.

Without this fix, for example, libffi might add /usr/lib/../lib to the
libs search directory, which is not filtered by ptxdist and will cause
compile errors on programs that link against libffi.

Signed-off-by: Michael Tretter <michael.tretter@posteo.net>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>