cross-binutils: fix broken symlinks for gcc-first
The symlinks pointed to non-existent files. As a result,
'<cross-prefix>-gcc -print-prog-name=objcopy' returned just 'objcopy' which
is the host objcopy and not the one from cross-binutils.

glibc-first uses this incorrect objcopy at build time. Building may fail
if the used objcopy is not multiarch capable.

With the correct symlinks -print-prog-name works correctly and produces an
absolute path to the correct objcopy.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>