replace host-mpfr with host-system-mpfr
This is a first step to make gcc plugins possible.
The header files for plugins include the headers for several of host
libraries that are currently statically compiled.
Use the system versions of those libraries. This way the user can install
the corresponding devel packages. Then gcc plugins can be built.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>