replace host-xz with host-system-xz
This is a first step to make gcc plugins possible.
The header files for plugins include the headers for several of host
libraries that are currently statically compiled.
Use the system versions of those libraries. This way the user can install
the corresponding devel packages. Then gcc plugins can be built.

xz (liblzma) is not needed for plugins but one more dependency into the
system makes no difference and this makes maintaining the toolchain easier.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>