don't used old icecc versions for cross compiling
To get a usefull target environment, we use icecc-create-env to package it,
then extract it and strip all binaries and then package it again.
Without the extra step the environment is huge, because the binaries are
not stripped at that point. However, if icecc-create-env compresses the
environment then all of this takes a long time.

And glibc 2.33 fails to build with old icecc versions.

So just avoid the whole issue on old distros and don't create a target
icecc environment.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>