rewrite PTXCONF_TOOLCHAIN_DEBUG and generall debug handling
The regular toolchains are already built with debug symbols, so that part
remains the same. PTXCONF_TOOLCHAIN_DEBUG just adds the sources for target
libraries to the toolchain.
Previously the source directory was changed but that makes the paths
longer, which can cause problems. So instead keep the directories and copy
the sources instead. Use -ffile-prefix-map= to rewrite the paths in the
debug symbols.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>