debian/rules: skip override_dh_update_autotools_config
During a regular build, this is a noop, because there are no autotools
files in the toolchain BSP.

During development, building packages may be restarted and in that case
override_dh_update_autotools_config will search and replace files in the
toolchain builds. That takes time and may change things compared to a
regular build.
So skip override_dh_update_autotools_config to avoid that.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>